+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Pagination of BIM360 issues
+- Pagination of issue comments, attachments, and root causes
 
 ## [8.0.2] - 2019-11-05